Code for the scoresheets GUI
I think this should work but something about the git connection wont let me run it? Put the sets(15) in their own gridpanes(5), then put those in the 'grid' in the right place.  Need to add the calculate button to the right place and make sure the textfields cant be edited.